+++ mayan/apps/locales/apps.py
@@ -2,7 +2,6 @@ import logging
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
-from django.contrib.auth.signals import user_logged_in
 from django.db.models.signals import post_save
 from django.utils.translation import ugettext_lazy as _
 
@@ -11,10 +10,7 @@ from mayan.apps.common.menus import menu_list_facet, menu_object
 from mayan.apps.events.classes import ModelEventType
 
 from .events import event_user_locale_profile_edited
-from .handlers import (
-    handler_user_locale_profile_session_config,
-    handler_user_locale_profile_create
-)
+from .handlers import handler_user_locale_profile_create
 from .links import (
     link_user_locale_profile_detail, link_user_locale_profile_edit
 )
@@ -59,9 +55,4 @@ class LocalesApp(MayanAppConfig):
             dispatch_uid='common_handler_user_locale_profile_create',
             receiver=handler_user_locale_profile_create,
             sender=settings.AUTH_USER_MODEL
-        )
-
-        user_logged_in.connect(
-            dispatch_uid='common_handler_user_locale_profile_session_config',
-            receiver=handler_user_locale_profile_session_config
         )